backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
google-cloud-vision
python-dotenv
scikit-learn
numpy
pandas
Pillow
python-multipart
pdf2image